[jOOQ/jOOQ#11323] Add HSQLDB procedural language support

+++ jOOQ/src/main/java/org/jooq/impl/Tools.java
@@ -5297,7 +5297,7 @@ final class Tools {
         if (updateCounts == null)
             throw new IllegalStateException();
         else
-            return updateCounts == 1;
+            return updateCounts.intValue() == 1;
     }
 
     /**
